SPT-1755: Ideas to rearrange code to make it clear
Ideas to rearrange code to make it clear
