--- src/commons/test-constants.ts
@@ -1,8 +1,0 @@
-export const TEST_SET_SIGNING_KEY_ID = "3a502a70-6b3b-4b65-bf74-23da88cc9a6";
-export const TEST_SET_SIGNING_KMS_KEY_ARN = `arn:aws:kms:eu-west-2:123456789012:key/${TEST_SET_SIGNING_KEY_ID}`;
-export const MOCKPKUNIT8ARRAY = new Uint8Array([
-  48, 89, 48, 19, 6, 7, 42, 134, 72, 206, 61, 2, 1, 6, 8, 42, 134, 72, 206, 61, 3, 1, 7, 3, 66, 0, 4, 125, 195, 233,
-  173, 119, 48, 219, 109, 73, 141, 131, 245, 12, 90, 157, 66, 160, 98, 97, 232, 110, 157, 17, 246, 247, 194, 149, 194,
-  241, 174, 157, 195, 245, 71, 202, 214, 134, 249, 40, 32, 34, 150, 110, 140, 161, 14, 235, 147, 152, 198, 239, 80, 49,
-  41, 97, 124, 140, 75, 231, 220, 84, 66, 158, 120,
-]);
