+++ src/commons/verifiable-credential-jwt.ts
