+++ src/handlers/ais-message-processor/tests/ais-message-handler.test.ts
@@ -10,7 +10,7 @@ import {
   GetLatestConfigurationCommandOutput,
 } from "@aws-sdk/client-appconfigdata";
 import { Uint8ArrayBlobAdapter } from "@smithy/util-stream";
-import { Configuration } from "../../../commons/configuration";
+import { Configuration } from "../../../services/configuration";
 import { MetricDimension, MetricName } from "../../../commons/metric-enum";
 import { AisMessage } from "../ais-message";
 import { MetricUnit } from "@aws-lambda-powertools/metrics";
