--- src/handlers/user-identity/user-identity-error-response.ts
@@ -1,4 +1,0 @@
-export type UserIdentityErrorResponse = {
-  error: string;
-  error_description: string;
-};
