--- src/handlers/user-identity/user-identity-request.ts
@@ -1,6 +1,0 @@
-import { IdentityVectorOfTrust } from "@govuk-one-login/data-vocab/credentials";
-
-export type UserIdentityRequest = {
-  vtr: IdentityVectorOfTrust[];
-  govukSigninJourneyId: string;
-};
