+++ src/services/tests/audit.test.ts
@@ -2,7 +2,7 @@ import { SendMessageCommand, SQSClient } from "@aws-sdk/client-sqs";
 import { mockClient } from "aws-sdk-client-mock";
 import "aws-sdk-client-mock-jest";
 import { auditIdentityRecordInvalidated } from "../audit";
-import { TxmaSisIdentityRecordInvalidated } from "../audit-events";
+import { TxmaSisIdentityRecordInvalidated } from "../../commons/audit-events";
 
 const sqsClientMock = mockClient(SQSClient);
 sqsClientMock.on(SendMessageCommand).resolves({});
